计算器从整数到浮点数的扩展以及增加了幂操作符运算

+++ stack_adt/include/calcuation_h/calcuation.h
@@ -30,7 +30,7 @@ char *infixToSuffix(const char *);
 // 根据后缀表达式计算结果
 // 前置条件：传入一个后缀表达式
 // 后置条件：返回这个后缀表达式计算出来的值
-int calcuationBySuffix(const char *);
+double calcuationBySuffix(const char *);
 
 // 读取用户输入的算式
 char *getUserInput(char *st, int n);
@@ -53,7 +53,7 @@ int getPrecedence(char ch);
 // 根据两个操作数和一个运算符进行运算
 // 前置条件：传入两个操作数和一个运算符
 // 后置条件：返回计算结果的整形
-int calcuation_two_nums(const char *num1, const char *num2, const char *oper);
+double calcuation_two_nums(const char *num1, const char *num2, const char *oper);
 
 // 系统运行函数
 void cal_sys_run(void);
